refactor: add isAnonymousFunction utility to check for anonymous functions
